+++ src/main/java/br/router/service/RouteService.java
@@ -1,5 +1,6 @@
 package br.router.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import br.router.model.Route;
@@ -7,7 +8,7 @@ import br.router.model.Vehicle;
 
 public interface RouteService {
 
-	Route getById(Long id);
+	Route getById(BigInteger id);
 	List<Route> getByVehicle(Vehicle vehicle);
 	List<Route> getAll();
 	Route save(Route route);
