+++ src/main/java/br/router/service/VehicleService.java
@@ -1,12 +1,13 @@
 package br.router.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import br.router.model.Vehicle;
 
 public interface VehicleService {
 
-	Vehicle getById(Long id);
+	Vehicle getById(BigInteger id);
 	List<Vehicle> getAll();
 	Vehicle save(Vehicle vehicle);
